added CascadeType ALL to delete compte,delete client works,update client works, package exception,message dto,localdate now 2

+++ src/main/java/com/example/demo/service/BankServiceImpl.java
@@ -6,6 +6,8 @@ import java.util.Optional;
 
 import org.springframework.stereotype.Service;
 
+import com.example.demo.exception.BankException;
+import com.example.demo.exception.NotFoundException;
 import com.example.demo.model.Client;
 import com.example.demo.repository.ClientRepository;
 
@@ -47,10 +49,12 @@ public class BankServiceImpl implements BankService {
 	@Override
 	public void deleteClient(Long id) //throws NotFoundException
 	{
-		/*if(!clientRepository.existsById(id))
-			throw new NotFoundException("account not found with id=" +id); */
-		//if solde === 0
-		 clientRepository.deleteById(id);
+		if(!clientRepository.existsById(id))
+			throw new NotFoundException("account not found with id=" +id); 
+	if(	clientRepository.findById(id).get().getCompteCourant().getInfos().getSolde() == 0.0) {
+		
+		clientRepository.deleteById(id);
+	}else throw new BankException("Pas de suppression possible car le compte a un solde positif ou négatif" +clientRepository.findById(id).get().getCompteCourant().getInfos().getSolde() )
 		
 	}
 }
